Display of student analytics in-line on the problem page.
Two problem types have graphical display:

  Radio button
  Checkbox

These other problem types display as in-line text:

  NumericalResponse
  StringResponse
  FormulaResponse

The data source is the analytics pipeline via the api.

The setting ANALYTICS_DATA_URL is used as an on/off switch for the feature.
The list of supported problem types is defined in INLINE_ANALYTICS_SUPPORTED_TYPES.

Conflicts:
	lms/djangoapps/courseware/module_render.py
	lms/djangoapps/courseware/tests/test_inline_analytics_integration.py
	lms/djangoapps/courseware/views.py
	lms/envs/aws.py
	lms/envs/common.py
	lms/static/js/inline_analytics.js
	lms/static/js/spec/main.js
	lms/templates/inline_analytics.html
	lms/urls.py
	requirements/edx/github.txt