Remove Image Modal HTML_Module template
This code has be refactored and revamped into a stand-alone XBlock [1].

These changes should have a significant effect on client-side
performance, as much of this content had been going across the wire on
every page request. Now, this content will only be loaded on courseware
pages which actually contain an Image Modal.

[1] https://github.com/Stanford-Online/xblock-factory/tree/image-modal