Limit Upload File Sizes to GridFS.
This commit puts a limit on the size of files that
course staff can upload to MongoDB. The limit is
enforced on the frontend in javascript as well
as backend via the /upload endpoint.

The limit is hard-coded in cms/envs/common.py and
may be changed according to the user's custom
needs.

If the user tries to upload a file that's too
large, an error message will pop up that points
the user to an external page with an alternate
upload procedure.

Also, adding my name onto list of contributors