Allow course creation after deletion at same URL
After deleting a course, creating one with the same URL did not
work because its mapping is not removed from the modulestore. This
fix adds a more robust check to see if a course with the URL
actually exists. The delete course functionality should ideally
update the mapping in the future.