Fix Multiple Choice Shuffle Bug
Multiple choice problems with the "shuffle" option set to "true" are now
shuffled on each page reload.

Previously, multiple choice problems "shuffled" into predictable patterns that
did not change after page reload or question reset.  The behavior and patterns
were consistent even when viewed by different users.  The problem was that the
seed used to randomize the order of the choices was never reset - once the
problem was created on studio, the ordering of the questions effectively
remained static.  Generating a new seed during each page reload fixes the
issue.

TODO:  Because the multiple choice shuffle bug involves the problem seed, the
shuffle bug is also related to issues with the randomization option.  There
have been issues in the past where the resetting of seeds with the
randomization option interferes with inline analytics.  Will need to check with
Backend in order to make sure that this fix does not cause the same problem.