Become teapot, short and stout
In-group, out-group, Shibboleth shout!

This seems to mostly work, so far.

I was able to follow through the code with some tracing statements on stage.
This revealed that attempting to sign in with a shib account did make it
most of the way through. It even generates a payload with redirect
information [1]. Sadly, this internal response is then dropped on the
floor [2].

From there, I tested on my devstack by hard-coding the aforementioned
payload [1]. As shib is not actually configured in development, I
couldn't get past the `/shib-login/` page, but I was routed there.

I'll need to put this on stage to verify the handoff to `/shib-login/`.
I don't know what, if any parameters may need passed along.

I did verify that attempting to enroll in a Shib-only course while
logged out will route you directly to web-auth and back again.

I have not tested registration via shib/web-auth.

[1] ```json
    {
        "redirect": "/shib-login/",
        "success": false
    }
    ```
[2] https://github.com/stvstnfrd/edx-platform/blob/shib-logistration/openedx/core/djangoapps/user_api/helpers.py#L401